feat: RecommendLive에서 사용하는 플레이어 상태를 훅으로 구현하여 상태 관심사 분리
